Update of Guzzle dependency & overall code clean up
