+++ .gitignore
@@ -1,3 +1,4 @@
 /vendor/
 /.idea/
 phpunit.xml
+composer.lock
