+++ phpunit.xml.dist
@@ -1,18 +1,13 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
-<!-- http://www.phpunit.de/manual/current/en/appendixes.configuration.html -->
-<phpunit
-    backupGlobals               = "false"
-    backupStaticAttributes      = "false"
-    colors                      = "true"
-    convertErrorsToExceptions   = "true"
-    convertNoticesToExceptions  = "true"
-    convertWarningsToExceptions = "true"
-    processIsolation            = "false"
-    stopOnFailure               = "false"
-    syntaxCheck                 = "false"
-    bootstrap                   = "./tests/bootstrap.php"
-    >
+<phpunit xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:noNamespaceSchemaLocation="http://schema.phpunit.de/4.1/phpunit.xsd"
+         backupGlobals="false"
+         colors="true"
+         bootstrap="./Tests/bootstrap.php"
+         failOnRisky="true"
+         failOnWarning="true"
+>
 
     <php>
         <env name="UNBABEL_USERNAME" value="*** YOUR API USERNAME HERE***" />
@@ -20,8 +15,8 @@
     </php>
 
     <testsuites>
-        <testsuite name="TestBundle Test Suite">
-            <directory>./tests/Unbabel/Tests</directory>
+        <testsuite name="Unbabel Test Suite">
+            <directory suffix="Test.php">./Tests</directory>
         </testsuite>
     </testsuites>
 
@@ -29,7 +24,7 @@
         <whitelist>
             <directory>./</directory>
             <exclude>
-                <directory>./tests</directory>
+                <directory>./Tests</directory>
                 <directory>./vendor</directory>
             </exclude>
         </whitelist>
